后端ssm--wxx

+++ ssm-web/src/main/java/com/hnguigu/controller/MenuController.java
@@ -1,6 +1,7 @@
 package com.hnguigu.controller;
 
 import com.hnguigu.service.MenuService;
+import com.hnguigu.service.RoleMenuService;
 import com.hnguigu.vo.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -16,6 +17,9 @@ public class MenuController {
 
     @Autowired
     MenuService menuService;
+
+    @Autowired
+    RoleMenuService roleMenuService;
 
     //通过条件查询所有
     @RequestMapping("/querymenucount.action")
@@ -76,16 +80,31 @@ public class MenuController {
 
         String[] idss=ids.split(",");
 
-        int num=menuService.deletePLMenu(idss);
+        int[] ints=new int[idss.length];
 
-        if(num==idss.length){
-            map.put("msg","删除成功");
-            map.put("code","1");
-        }else{
-            map.put("msg","删除失败");
-            map.put("code","0");
+        int count=0;
+        for(int i=0;i<idss.length;i++) {
+            ints[i] = Integer.parseInt(idss[i]);
+
+            //判断是否有某个菜单被角色引用，如果有不能删除
+            count = roleMenuService.queryRMMid(ints[i]);
+
         }
 
+        if(count>0){
+            map.put("msg","对不起，其中有菜单已被引用，不能删除");
+            map.put("code","2");
+        }else {
+            int num=menuService.deletePLMenu(idss);
+
+            if(num==idss.length){
+                map.put("msg","删除成功");
+                map.put("code","1");
+            }else{
+                map.put("msg","删除失败");
+                map.put("code","0");
+            }
+        }
 
         return map;
     }
@@ -96,15 +115,25 @@ public class MenuController {
     @CrossOrigin
     public Map<String,String> deleteMenu(int id){
         Map<String,String> map=new HashMap<>();
-        int num=menuService.deleteMenu(id);
 
-        if(num==1){
-            map.put("msg","删除成功");
-            map.put("code","1");
-        }else{
-            map.put("msg","删除失败");
-            map.put("code","0");
+        //判断是否有某个菜单被角色引用，如果有不能删除
+        int count=roleMenuService.queryRMMid(id);
+
+        if(count>0){
+            map.put("msg","对不起，该菜单已被引用，不能删除");
+            map.put("code","2");
+        }else {
+            int num=menuService.deleteMenu(id);
+
+            if(num==1){
+                map.put("msg","删除成功");
+                map.put("code","1");
+            }else{
+                map.put("msg","删除失败");
+                map.put("code","0");
+            }
         }
+
         return map;
     }
